% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{activity_properties}
\alias{activity_properties}
\alias{nr_activities}
\alias{has_any_activity}
\alias{any_activity}
\alias{get_activity}
\title{Activity Properties}
\description{
\itemize{
\item activities: Return activities in a schedule as data frame.
\item DESCREVER CADA UMA DAS COLUNAS DO DATA FRAME
\item $get_activity(id)
\item nr_activities: Number of activities in a schedule as an integer value.
\item has_any_activity: A logical value that indicate if the schedule has any activity.
}
}
