% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_act_rel.R
\name{add_act_rel}
\alias{add_act_rel}
\title{Add an activity and her relations}
\usage{
add_act_rel(schedule, id, name, duration, relations_id=c(), dir="succ")
}
\arguments{
\item{schedule}{Schedule in which the activity and the relations will be added.}

\item{id}{Activity id. The id will be used to make relation between activities.}

\item{name}{The name of activity.}

\item{duration}{A number that represents the activity's duration.
It must be equal or grater than zero.}

\item{relations_id}{A vector of ids such that will be linked with activity id.
It may be a relations of successor or predecessors.}

\item{dir}{Direction of relations_id: It may be "succ" or "pred".
If dir="succ" the relations_id will be the successor of tye activity.
If dir="pred" the relations_id will be the predecessor of the activity.}
}
\value{
schedule with an activity and relations added to it.
}
\description{
Add an activity and her relations to a schedule.
}
\examples{
schedule <- make_empty_schedule(
  "Fictitous Project Example",
  "VANHOUCKE, Mario. Measuring time:
    improving project performance using earned value management.
    Gent: Springer, 2009, p. 18"
)
schedule <- add_act_rel(schedule,  1, "a1" , 0, c(2,3,4))
schedule <- add_act_rel(schedule,  2, "a2" , 4, c(5))
schedule <- add_act_rel(schedule,  3, "a3" , 9, c(10))
schedule <- add_act_rel(schedule,  4, "a4" , 1, c(6))
schedule <- add_act_rel(schedule,  5, "a5" , 4, c(9))
schedule <- add_act_rel(schedule,  6, "a6" , 5, c(7))
schedule <- add_act_rel(schedule,  7, "a7" , 1, c(8,11))
schedule <- add_act_rel(schedule,  8, "a8" , 7, c(12))
schedule <- add_act_rel(schedule,  9, "a9" , 8, c(12))
schedule <- add_act_rel(schedule, 10, "a10", 3, c(12))
schedule <- add_act_rel(schedule, 11, "a11", 3, c(12))
schedule <- add_act_rel(schedule, 12, "a12", 0)
message("Schedule info:")
schedule$info
message("Activities info:")
schedule$activities

}
\references{
DOCUMENTAÇÃO SOBRE O MÉTODO: livros, artigos ou sites.
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
