% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_activity.R
\name{add_activity}
\alias{add_activity}
\title{Add an activity to a schedule}
\usage{
add_activity(schedule, id, name="", duration=0)
}
\arguments{
\item{schedule}{Schedule in which the activity will be added.}

\item{id}{Activity id. The id will be used to make relation between activities.}

\item{name}{The name of activity. The default is a empty string.}

\item{duration}{A number that represents the activity's duration.
It must be equal or grater than zero. The default value is zero.}
}
\value{
A Schedule object with an activity added and critical path calculated.
}
\description{
Add an activity to a schedule that must exist.
}
\examples{
# Create an empty schedule
sch <- make_empty_schedule("A Project", "From criticalpath package")
# Add activities to it
sch <- add_activity(sch, 1, "A", 5)
sch <- add_activity(sch, 2, "B", 4)
sch <- add_activity(sch, 3, "C", 2)
sch <- add_activity(sch, 4, "D", 3)
sch <- add_activity(sch, 5, "E", 5)
sch <- add_activity(sch, 6, "F", 3)
# Get schedule info
# What is the project duration?
message(paste("Project duration:", sch$info$duration))
# Which activities are critical?
sch$activities$name[sch$activities$critical]


}
\references{
Csardi, G. & Nepusz, T. (2005).
The Igraph Software Package for Complex Network Research.
\emph{InterJournal}. Complex Systems. 1695.
\href{https://www.researchgate.net/publication/221995787_The_Igraph_Software_Package_for_Complex_Network_Research}{Article}
/ \href{https://igraph.org/}{igraph}

Vanhoucke, M. (2009) \emph{Measuring Time}:
Improving Project Performance Using Earned Value Management.
Springer-Verlag US.
doi: \href{https://doi.org/10.1007/978-1-4419-1014-1}{10.1007/978-1-4419-1014-1}.

Vanhoucke, M. (2013) \emph{Project Management with Dynamic Scheduling}:
Baseline Scheduling, Risk Analysis and Project Control.
Springer-Verlag Berlin Heidelberg.
doi: \href{https://doi.org/10.1007/978-3-642-40438-2}{10.1007/978-3-642-40438-2}

Vanhoucke, M. (2014) \emph{Integrated Project Management and Control}:
First Comes the Theory, then the Practice.
Springer International Publishing Switzerland.
doi: \href{https://doi.org/10.1007/978-3-319-04331-9}{10.1007/978-3-319-04331-9}
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
