% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_relation.R
\name{add_relation}
\alias{add_relation}
\title{Add a relation to a schedule.}
\usage{
add_relation(schedule, from_id, to_id, type, lag)
}
\arguments{
\item{schedule}{Schedule in which the relation will be added.}

\item{from_id}{The id of predecessor activity.}

\item{to_id}{The id of successor activity.}

\item{type}{Specifies the type of relation between activities.
Its value may be: FS, FF, SS, SF, that means:

FS: Finish-Start relation.
Activity to_id can only start after the finish of activity from_id.

FF: Finish-Finish relation.
Activity to_id must finish together with activity from_id.

SS: Start-Start relation.
Activity to_id must start together with activity from_id.

SF: Start-Finish relation.
Activity to_id must finish when activity from_id starts.}

\item{lag}{The time period between activities that the successor activity
must be advanced, or lated, after activity from_id.
The must be a integer, less than, equal or greater than zero.}
}
\value{
A Schedule object with a relation added and critical path calculated.
}
\description{
Add a relation to a schedule.
If type is not defined, it is assumed to be FS.
If lag is not defined, it is assumed to be zero.
}
\examples{
# Create an empty schedule
sch <- make_empty_schedule("A Project", "From criticalpath package")
# Add activities to it
sch <- add_activity(sch, 1, "A", 5)
sch <- add_activity(sch, 2, "B", 4)
sch <- add_activity(sch, 3, "C", 2)
sch <- add_activity(sch, 4, "D", 3)
sch <- add_activity(sch, 5, "E", 4)
sch <- add_activity(sch, 6, "F", 3)
# Add relations between activities
sch <- add_relation(sch, 1, 2)
sch <- add_relation(sch, 1, 3)
sch <- add_relation(sch, 2, 4)
sch <- add_relation(sch, 3, 5)
sch <- add_relation(sch, 4, 6)
sch <- add_relation(sch, 5, 6)
# Get schedule info
# What is the project duration?
message(paste("Project duration:", sch$info$duration))
# Which activities are critical?
sch$activities$name[sch$activities$critical]

}
\references{
Csardi, G. & Nepusz, T. (2005).
The Igraph Software Package for Complex Network Research.
\emph{InterJournal}. Complex Systems. 1695.
\href{https://www.researchgate.net/publication/221995787_The_Igraph_Software_Package_for_Complex_Network_Research}{Article}
/ \href{https://igraph.org/}{igraph}

Vanhoucke, M. (2009) \emph{Measuring Time}:
Improving Project Performance Using Earned Value Management.
Springer-Verlag US.
doi: \href{https://doi.org/10.1007/978-1-4419-1014-1}{10.1007/978-1-4419-1014-1}.

Vanhoucke, M. (2013) \emph{Project Management with Dynamic Scheduling}:
Baseline Scheduling, Risk Analysis and Project Control.
Springer-Verlag Berlin Heidelberg.
doi: \href{https://doi.org/10.1007/978-3-642-40438-2}{10.1007/978-3-642-40438-2}

Vanhoucke, M. (2014) \emph{Integrated Project Management and Control}:
First Comes the Theory, then the Practice.
Springer International Publishing Switzerland.
doi: \href{https://doi.org/10.1007/978-3-319-04331-9}{10.1007/978-3-319-04331-9}
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
