% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_predecessors.R
\name{all_predecessors}
\alias{all_predecessors}
\title{List All Predecessors}
\usage{
all_predecessors(schedule, id, ign_from)
}
\arguments{
\item{schedule}{A schedule}

\item{id}{Activity id to listed}

\item{ign_from}{An arc to be ignored: ign_from -> id}
}
\value{
A vector
}
\description{
List all predecessors from an activity: direct or indirect predecessors.
}
\examples{
all_predecessors(schedule, 5)

}
\references{
DOCUMENTAÇÃO SOBRE O MÉTODO: livros, artigos ou sites.
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
