% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_schedule_status.R
\name{cpt_is_schedule_created}
\alias{cpt_is_schedule_created}
\title{Is the schedule created?}
\usage{
cpt_is_schedule_created(sch)
}
\arguments{
\item{sch}{A schedule of class \code{KKKKKKKK}}
}
\value{
A logical value.
}
\description{
Is the schedule created?
}
\examples{

}
