% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_schedule_status.R
\name{cpt_is_schedule_empty}
\alias{cpt_is_schedule_empty}
\title{Is the schedule empty?}
\usage{
cpt_is_schedule_empty(sch)
}
\arguments{
\item{sch}{A schedule of class \code{KKKKKKKK}}
}
\value{
A logical value.
}
\description{
Is the schedule empty?
}
\examples{

}
