% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_schedule_status.R
\name{cpt_is_schedule_with_some_activity}
\alias{cpt_is_schedule_with_some_activity}
\title{Is schedule with some activity?}
\usage{
cpt_is_schedule_with_some_activity(sch)
}
\arguments{
\item{sch}{A schedule of class \code{KKKKKKKK}}
}
\value{
A logical value.
}
\description{
Is schedule with some activity?
}
\examples{

}
