% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_topological_organization.R
\name{cpt_topological_organization}
\alias{cpt_topological_organization}
\title{FALAR QUE "A ORDEM DOS RELACIONAMENTOS MUDA
PARA FICAR DE ACORDO COM A ORDENAÇÃO TOPOLÓGICA".}
\usage{
cpt_topological_organization(sch)
}
\description{
Para voltar na ordem original, basta utilizar a propriedade \verb{$ord}.
}
