% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{from_data_frame}
\alias{from_data_frame}
\alias{schedule_from_data_frame}
\alias{as_data_frame}
\alias{relations_as_data_frame}
\title{Create a schedule object from data frame or vice-versa}
\arguments{
\item{title}{A project title for identification.}

\item{reference}{A reference from project origin,
for example, a book, a paper, a corporation, or nothing.}

\item{activities}{data frame with activities.}

\item{relations}{data frame with precedence relations between activities.}
}
\value{
A Schedule object with an activities and relations.
Or a data frame for activities or relations.
}
\description{
The method Schedule$new(title, reference, activities, relations)
creates an schedule object  from one or two data frames,
one containing activities lists and the order the precedence relations
between activities. After creation, it is applied the Critical Path Method (CPM).

The method Schedule$activities and Schedule$relations
does the opposite: it returns the activities data frame
and the relation data frame from a schedule.
}
\examples{
activities <- data.frame(
  id        = 1:17,
  name      = paste("a", as.character(1:17), sep=""),
  duration  = c(1,2,2,4,3,3,3,2,1,1,2,1,1,1,1,2,1)
)

relations <- data.frame(
  from = c(1, 1, 2, 2, 2, 3, 3, 3,  3,  4,  5,  6,
           7,  8,  9, 10, 11, 11, 12, 12, 13, 13, 14, 14, 15, 15),
  to   = c(2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 11, 11,
           12, 13, 14, 15, 16, 17, 16, 17, 16, 17, 16, 17, 16, 17)
)
schedule <- Schedule$new(activities, relations)
schedule$title <- "Project 1: Cost Information System"
schedule$reference <- "VANHOUCKE, Mario.
Integrated project management and control:
  first comes the theory, then the practice.
  Gent: Springer, 2014, p. 6"
}
\seealso{
\link{criticalpath}, \link{Schedule}.
}
\author{
Rubens Jose Rosa (rubens@rubensjoserosa.com),
Marcos dos Santos, Thiago Marques
}
