% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt_matrix.R
\name{gantt_matrix}
\alias{gantt_matrix}
\title{Gantt Matrix}
\usage{
gantt_matrix(schedule)
}
\arguments{
\item{schedule}{Schedule to crate Gantt mattix.}
}
\value{
Uma matriz, onde 1 indicada que a atividade esta em execucao
Nessa matriz as linhas representam as atividades
As colunas representam o periodo de execucao da atividade.

A matrix, where row represent activity and columns representa time period.
A cell represents a activity in execution in period:  1 indicate that a activity is in execution.
}
\description{
Create a matrix that representa a Gantt chart.
}
\examples{
activities <- data.frame(
  id        = 1:17,
  name      = paste("a", as.character(1:17), sep=""),
  duration  = c(1,1,3,2, 2,2,2,1, 4,5,3,3, 4,5,1,5,2)
)
relations <- data.frame(
  from = c(1, 2, 3, 3, 4, 5, 6, 7, 8,  8,  8,  8,  8,  9,
           10, 11, 12, 13, 13, 14, 14, 15, 15),
  to   = c(2, 3, 4, 6, 5, 8, 7, 8, 9, 10, 11, 12, 13, 14,
           14, 14, 14, 14, 15, 16, 17, 16, 17)
)
schedule <- schedule_from_data_frame(
  activities,
  relations,
  "Project 2: Patient Transport System",
  "VANHOUCKE, Mario. Integrated project management and control:
    first comes the theory, then the practice. Gent: Springer, 2014, p. 9"
)
gantt <- gantt_matrix(schedule)
effort_by_period <- colSums(gantt)
s_curve <- cumsum(effort_by_period)
#plot(s_curve, type="l", lwd=3)

}
\references{
DOCUMENTAÇÃO SOBRE O MÉTODO: livros, artigos ou sites.
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
