% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_redundant.R
\name{is_redundant}
\alias{is_redundant}
\title{Is an arc redundant}
\usage{
is_redundant(schedule, id_from, id_to)
}
\arguments{
\item{schedule}{A Schedule.}

\item{id_from}{Activity id from.}

\item{id_to}{Activity id to.}
}
\value{
A logical TRUE if an arc is redundant; FALSE if not.
}
\description{
Verifify if an arc between two activities, id_from -> id_to, is redundant.
}
\examples{
is_redundant(schedule, 1, 8)

}
\references{
DOCUMENTAÇÃO SOBRE O MÉTODO: livros, artigos ou sites.
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
