% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_empty_schedule.R
\name{make_empty_schedule}
\alias{make_empty_schedule}
\title{Make an empty schedule}
\usage{
make_empty_schedule(title="", reference="")
}
\arguments{
\item{title}{A project title for identification.}

\item{reference}{A reference from project origin,
for example, a book, a paper, a corporation, or nothing.}
}
\value{
A Schedule object with no activities and no relations.
}
\description{
Make a schedule with no activities and no relations.
The activities and relations will be add in sequence.
}
\examples{
schedule <- make_empty_schedule("P1", "From criticalpath package")
}
