% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{nr_activities}
\alias{nr_activities}
\alias{activities}
\alias{has_any_activity}
\alias{any_activity}
\title{Number of  activities / Has any activity}
\value{
An integer indicating the number of activities
or a logical value TRUE, meaning that the schedule has any activity
or FALSE, for a empty schedule.
}
\description{
Number of activities in a schedule as an integer value,
or a logical value that indicate if the schedule has any activity.
}
\examples{
schedule <- Schedule$new()
schedule$has_any_activity # FALSE
schedule$nr_activities # 0

schedule <- Schedule$new()
schedule$add_act_rel(1, "A" , 3, c(2,3))
schedule$add_act_rel(2, "B" , 4, c(4))
schedule$add_act_rel(3, "C" , 7, c(4))
schedule$add_act_rel(4, "D" , 5)
schedule$has_any_activity # TRUE
schedule$nr_activities # 4
}
\seealso{
\link{criticalpath}, \link{Schedule}.
}
\author{
Rubens Jose Rosa (rubens@rubensjoserosa.com),
Marcos dos Santos, Thiago Marques
}
