% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{nr_relations}
\alias{nr_relations}
\alias{relations}
\alias{has_any_relations}
\alias{any_relations}
\title{Number of  relations / Has any relation}
\value{
An integer indicating the number of relations
or a logical value TRUE, meaning that the schedule has any activity
or FALSE, for a empty schedule.
}
\description{
Number of relations (arcs) in a schedule as an integer value,
or a logical value that indicates if the schedule has any relation.
}
\seealso{
\link{criticalpath}, \link{Schedule}.
}
\author{
Rubens Jose Rosa (rubens@rubensjoserosa.com), Marcos dos Santos, Thiago Marques
}
