% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Pipe assign operator}
\usage{
lhs \%<>\% rhs
}
\arguments{
\item{lhs}{A variable.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
The result of calling \code{lhs <- rhs(lhs)}.
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%<>\%}} for details.
}
\keyword{internal}
