% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{reference}
\alias{reference}
\alias{title}
\alias{duration}
\alias{schedule_duration}
\title{Title, Reference and Duration}
\description{
\strong{title} A project title for identification.
\strong{reference} A reference from project origin,
for example, a book, a paper, a corporation, or nothing.
\itemize{
\item title: A project title for identification.
\item reference: A reference from project origin, for example, a book, a paper, a corporation, or nothing.
\item duration: An integer value that indicate the project duration calculated by CPM.
An integer value that indicate the duration of a schedule.
}
}
