% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{relation_properties}
\alias{relation_properties}
\alias{nr_relations}
\alias{has_any_relation}
\alias{any_relation}
\title{Relation Properties}
\description{
\itemize{
\item relations: Return relations in a schedule as data frame.
\item DESCREVER CADA UMA DAS COLUNAS DO DATA FRAME
\item nr_relations: Number of relations (arcs) in a schedule as an integer value.
\item has_any_relation: A logical value that indicate if the schedule has any relation.
}
}
