% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch_activity.R
\name{sch_add_activities_tibble}
\alias{sch_add_activities_tibble}
\title{Add Activities Tibble}
\usage{
sch_add_activities_tibble(sch, atb)
}
\arguments{
\item{sch}{A schedule object.}

\item{atb}{A tibble with activities definitions.}
}
\value{
A schedule with a activity tibble (atb) added.
}
\description{
Add activities tibble to a schedule.
}
\examples{
atb <- tibble::tibble(
  id        = 1:17,
  name      = paste("a", as.character(1:17), sep=""),
  duration  = c(1L,2L,2L,4L,3L,3L,3L,2L,1L,1L,2L,1L,1L,1L,1L,2L,1L)
)
sch <- sch_new() \%>\%
  sch_add_activities_tibble(atb) \%>\%
  sch_plan()
sch_duration(sch) #4
# sch_activities(sch)

}
