% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch_relation.R
\name{sch_is_redundant}
\alias{sch_is_redundant}
\title{Is Redundant}
\usage{
sch_is_redundant(sch, id_from, id_to)
}
\arguments{
\item{sch}{A schedule object.}

\item{id_from}{From activity id.}

\item{id_to}{To activity id.}
}
\value{
A logical \code{TRUE} if an arc is redundant;
\code{FALSE} if it is not.
}
\description{
Verify if a relation between two activities is redundant.
A relation A->C is redundant if there are A->C, A->B, B->C relations.
}
\examples{
sch <- sch_new() \%>\%
  sch_title("Fictitious Project Example") \%>\%
  sch_reference("VANHOUCKE, Mario. Measuring time:
  improving project performance using earned value management.
  Gent: Springer, 2009, p. 18") \%>\%
  sch_add_activity( 2L, "a2" , 4L,  5L, 12L) \%>\%
  sch_add_activity( 3L, "a3" , 9L, 10L) \%>\%
  sch_add_activity( 4L, "a4" , 1L,  6L) \%>\%
  sch_add_activity( 5L, "a5" , 4L,  9L) \%>\%
  sch_add_activity( 6L, "a6" , 5L,  7L) \%>\%
  sch_add_activity( 7L, "a7" , 1L,  8L,11L) \%>\%
  sch_add_activity( 8L, "a8" , 7L, 12L) \%>\%
  sch_add_activity( 9L, "a9" , 8L, 12L) \%>\%
  sch_add_activity(10L, "a10", 3L, 12L) \%>\%
  sch_add_activity(11L, "a11", 3L, 12L) \%>\%
  sch_add_activity(12L, "a12", 0L) \%>\%
  sch_plan()
sch_is_redundant(sch, 2, 5) # FALSE
sch_is_redundant(sch, 2, 12) # TRUE

}
\seealso{
\code{\link[=sch_all_predecessors]{sch_all_predecessors()}}, \code{\link[=sch_all_successors]{sch_all_successors()}},
\code{\link[=sch_gantt_matrix]{sch_gantt_matrix()}}, \code{\link[=sch_relations]{sch_relations()}}, \code{\link[=sch_predecessors]{sch_predecessors()}},
\code{\link[=sch_successors]{sch_successors()}}, \code{\link[=sch_activities]{sch_activities()}}.
}
