% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schedule.R
\name{Schedule}
\alias{Schedule}
\title{R6 Class Representing a Schedule}
\usage{
add_activity(schedule, id, name="", duration=0)
}
\description{
A schedule has activities and relations data-frames.
}
\details{
With a Schedule, you can calculate the project critical path.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Schedule$new()}}
\item \href{#method-add_activity}{\code{Schedule$add_activity()}}
\item \href{#method-add_relation}{\code{Schedule$add_relation()}}
\item \href{#method-add_act_rel}{\code{Schedule$add_act_rel()}}
\item \href{#method-print}{\code{Schedule$print()}}
\item \href{#method-activities_as_data_frame}{\code{Schedule$activities_as_data_frame()}}
\item \href{#method-all_successors}{\code{Schedule$all_successors()}}
\item \href{#method-all_predecessors}{\code{Schedule$all_predecessors()}}
\item \href{#method-change_durations}{\code{Schedule$change_durations()}}
\item \href{#method-get_activity}{\code{Schedule$get_activity()}}
\item \href{#method-gantt_matrix}{\code{Schedule$gantt_matrix()}}
\item \href{#method-xy_gantt_matrix}{\code{Schedule$xy_gantt_matrix()}}
\item \href{#method-is_redundant}{\code{Schedule$is_redundant()}}
\item \href{#method-topoi_sp}{\code{Schedule$topoi_sp()}}
\item \href{#method-topoi_ad}{\code{Schedule$topoi_ad()}}
\item \href{#method-topoi_la}{\code{Schedule$topoi_la()}}
\item \href{#method-topoi_tf}{\code{Schedule$topoi_tf()}}
\item \href{#method-clone}{\code{Schedule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$new(title = "", reference = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title ...}

\item{\code{reference}}{reference ...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An Schedule object.
Add an activity to a schedule

Add an activity to a schedule that must exist.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_activity"></a>}}
\if{latex}{\out{\hypertarget{method-add_activity}{}}}
\subsection{Method \code{add_activity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$add_activity(id, name = "", duration = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Activity id. The id will be used to make relation between activities.}

\item{\code{name}}{The name of activity. The default is a empty string.}

\item{\code{duration}}{A number that represents the activity's duration.
It must be equal or grater than zero. The default value is zero.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Schedule object with an activity added and critical path calculated.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_relation"></a>}}
\if{latex}{\out{\hypertarget{method-add_relation}{}}}
\subsection{Method \code{add_relation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$add_relation(from_id, to_id, type = "FS", lag = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_act_rel"></a>}}
\if{latex}{\out{\hypertarget{method-add_act_rel}{}}}
\subsection{Method \code{add_act_rel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$add_act_rel(id, name, duration, relations_id = c(), dir = "succ")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-activities_as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-activities_as_data_frame}{}}}
\subsection{Method \code{activities_as_data_frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$activities_as_data_frame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_successors"></a>}}
\if{latex}{\out{\hypertarget{method-all_successors}{}}}
\subsection{Method \code{all_successors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$all_successors(id, ign_to = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_predecessors"></a>}}
\if{latex}{\out{\hypertarget{method-all_predecessors}{}}}
\subsection{Method \code{all_predecessors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$all_predecessors(id, ign_from = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-change_durations"></a>}}
\if{latex}{\out{\hypertarget{method-change_durations}{}}}
\subsection{Method \code{change_durations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$change_durations(new_durations)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_activity"></a>}}
\if{latex}{\out{\hypertarget{method-get_activity}{}}}
\subsection{Method \code{get_activity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$get_activity(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gantt_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-gantt_matrix}{}}}
\subsection{Method \code{gantt_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$gantt_matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xy_gantt_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-xy_gantt_matrix}{}}}
\subsection{Method \code{xy_gantt_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$xy_gantt_matrix(gantt = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_redundant"></a>}}
\if{latex}{\out{\hypertarget{method-is_redundant}{}}}
\subsection{Method \code{is_redundant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$is_redundant(id_from, id_to)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topoi_sp"></a>}}
\if{latex}{\out{\hypertarget{method-topoi_sp}{}}}
\subsection{Method \code{topoi_sp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$topoi_sp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topoi_ad"></a>}}
\if{latex}{\out{\hypertarget{method-topoi_ad}{}}}
\subsection{Method \code{topoi_ad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$topoi_ad()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topoi_la"></a>}}
\if{latex}{\out{\hypertarget{method-topoi_la}{}}}
\subsection{Method \code{topoi_la()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$topoi_la()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topoi_tf"></a>}}
\if{latex}{\out{\hypertarget{method-topoi_tf}{}}}
\subsection{Method \code{topoi_tf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$topoi_tf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
