% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{schedule_duration}
\alias{schedule_duration}
\alias{duration}
\title{Duration of a schedule}
\value{
An integer indicating the duration of of a schedule.
}
\description{
An integer value that indicate the duration of a schedule.
}
\seealso{
\link{criticalpath}, \link{Schedule}.
}
\author{
Rubens Jose Rosa (rubens@rubensjoserosa.com), Marcos dos Santos, Thiago Marques
}
