% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{schedule_from_act_rel}
\alias{schedule_from_act_rel}
\alias{add_act_rel}
\title{Schedule with activities and relations added together}
\description{
\itemize{
\item $add_act_rel(id, name, duration, relations_id=c(), direction="succ")
}
}
