% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{schedule_from_activities}
\alias{schedule_from_activities}
\alias{add_activity}
\alias{add_activities}
\title{Add activities to a schedule.}
\arguments{
\item{id}{Activity id, an integer number that
must be unique within a schedule.
It will be used to make relation between activities.}

\item{name}{The name of activity. The default is an empty string.}

\item{duration}{A number that represents the activity's duration.
It is integer number without time unit.
It must be equal or grater than zero. The default value is zero.}

\item{activities}{Data frame with activities.
If it is not informed, the schedule will be created without any activity.
Its structure is defined in \link{schedule_from_data_frame}}
}
\value{
A Schedule object with an activity added and
the critical path calculated.
}
\description{
\code{Schedule$add_activity(id, name, duration)}
adds an activity to a schedule.

\code{Schedule$add_activities(activities)}
adds an activities data frame to a schedule.

Usage

Schedule$add_activity(id, name, duration)

Schedule$add_activities(activities)
}
\examples{
# Activities added one by one
schedule <- Schedule$new()
schedule$add_activity(1, "Task 1", 5)
schedule$add_activity(2, "Task 2", 6)
schedule$add_activity(3, "Task 3", 8)
schedule$add_activity(4, "Task 4", 6)
schedule$add_activity(5, "Task 5", 9)
schedule$add_activity(6, "Task 6", 3)
schedule$add_activity(7, "Task 7", 4)

# Activities add once for all
activities <- data.frame(
  id        = 1:17,
  name      = paste("a", as.character(1:17), sep=""),
  duration  = c(1,2,2,4,3,3,3,2,1,1,2,1,1,1,1,2,1)
)
schedule <- Schedule$new()
schedule$add_activities(activities)

}
\seealso{
\link{schedule_from_data_frame}, \link{add_relation}, \link{criticalpath}, \link{Schedule}.
}
\author{
Rubens Jose Rosa (rubens@rubensjoserosa.com),
Marcos dos Santos, Thiago Marques
}
