% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_from_data_frame.R
\name{schedule_from_data_frame}
\alias{schedule_from_data_frame}
\title{Create a schedule from data frame}
\usage{
schedule_from_data_frame(activities, relations, title, reference)
}
\arguments{
\item{activities}{data frame with activities.}

\item{relations}{data frame with precedence relations between activities.}

\item{title}{A project title for identification.}

\item{reference}{A reference from project origin,
for example, a book, a paper, a corporation, or nothing.}
}
\value{
A Schedule object with an activities and relations.
}
\description{
Create a schedule from data frames and apply Critical Path Method (CPM)
}
