% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{schedule_from_relations}
\alias{schedule_from_relations}
\alias{add_relation}
\alias{add_relations}
\title{Schedule with relations added one by one}
\description{
\itemize{
\item $add_relation(from_id, to_id, type="FS", lag=0)
\item $add_relations(relations)
}
}
