% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{sucessors}
\alias{sucessors}
\alias{all_successors}
\alias{all_predecessors}
\alias{predecessors}
\alias{is_redundant}
\alias{redundant}
\title{Successors and Predecessors}
\description{
\itemize{
\item $all_successors(id, ign_to=NULL)
\item $all_predecessors(id, ign_from=NULL)
\item $is_redundant(id_from, id_to)
}
}
