% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoi_sp.R
\name{topoi_sp}
\alias{topoi_sp}
\title{SP Serial or Parallel Topological Indicator}
\usage{
topoi_sp(schedule)
}
\arguments{
\item{schedule}{A Schedule object.}
}
\value{
A number between 0 and 1, inclusive.
}
\description{
It shows the closeness of a network to a serial or parallel graph (1, 20).
}
\details{
As the network becomes serial, the SP increase, until one,
when the network totaly serial
}
\examples{
schedule <- make_empty_schedule("P2", "From criticalpath package")
schedule <- add_act_rel(schedule, 1, "A", 5, c(2,3))
schedule <- add_act_rel(schedule, 2, "B", 4, c(4))
schedule <- add_act_rel(schedule, 3, "C", 2, c(5))
schedule <- add_act_rel(schedule, 4, "D", 3, c(6))
schedule <- add_act_rel(schedule, 5, "E", 4, c(6))
schedule <- add_act_rel(schedule, 6, "F", 3)
topoi_sp(schedule)

}
\references{
DOCUMENTAÇÃO SOBRE O MÉTODO: livros, artigos ou sites.
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
