% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_gantt_matrix.R
\name{xy_gantt_matrix}
\alias{xy_gantt_matrix}
\title{Transform a Gantt matrix in x,y coordinates}
\usage{
xy_gantt_matrix(ganttm)
}
\arguments{
\item{ganttm}{A Gantt Matrix.}
}
\value{
A matrix x, y and weight.
}
\description{
Transform a Gantt matrix in x,y coordinates.
Each point greater than zero in a Gantt matrix becomes a x, y coordinate.
}
\examples{
activities <- data.frame(
  id        = 1:17,
  name      = paste("a", as.character(1:17), sep=""),
  duration  = c(1,1,3,2, 2,2,2,1, 4,5,3,3, 4,5,1,5,2)
)
relations <- data.frame(
  from = c(1, 2, 3, 3, 4, 5, 6, 7, 8,  8,  8,  8,  8,  9, 10,
           11, 12, 13, 13, 14, 14, 15, 15),
  to   = c(2, 3, 4, 6, 5, 8, 7, 8, 9, 10, 11, 12, 13, 14, 14,
           14, 14, 14, 15, 16, 17, 16, 17)
)
schedule <- schedule_from_data_frame(
  activities,
  relations,
  "Project 2: Patient Transport System",
  "VANHOUCKE, Mario. Integrated project management and control:
    first comes the theory, then the practice. Gent: Springer, 2014, p. 9"
)
gantt <- gantt_matrix(schedule)
xygantt <- xy_gantt_matrix(gantt)
#plot(xygantt[,1], -xygantt[,2])

}
\references{
DOCUMENTAÇÃO SOBRE O MÉTODO: livros, artigos ou sites.
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
