% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{activity_properties}
\alias{activity_properties}
\alias{nr_activities}
\alias{has_any_activity}
\alias{any_activity}
\alias{activities}
\alias{get_activity}
\title{Activity Properties}
\description{
There several methods about activities that you can use to get
information about them. After you add activities to a schedule, you may
want to know how much activities has in a schedule or when each activity
is planned to start or to finish.
In this section, we will explain how you can find these information.
\itemize{
\item \strong{Has Any Activity:} A logical value that indicate if the schedule
has any activity. A TRUE value means that the schedule has some
activity; a FALSE, means that the schedule is empty.
\itemize{
\item Usage: \code{Schedule$has_any_activity}
}
\item \strong{Number of Activities:} Number of activities in a schedule
as an integer value.
\itemize{
\item Usage: \code{Schedule$nr_activities}
}
\item \strong{Get activity:} Gets an activity by id. It returns a data
frame with one line about activity. The structure of a data is
described in next topic.
\itemize{
\item Usage: \code{Schedule$get_activity(id)}
}
\item \strong{Activities:} Return a data frame with all activities of a schedule
in a activity id order. This is the main information calculated by CPM.
The data frame is formed by following structure:
\itemize{
\item \strong{id:} Activity id.
\item \strong{name:} The name of activity.
\item \strong{duration:} A number that represents the activity's duration.
\item \strong{milestone:} A milestone is an activity with zero duration.
This property indicates if a activity is a milestone or not:
\code{TRUE} indicates it is a milestone; \code{FALSE} indicates it is not.
\item \strong{critical:} A critical activity is one with total float minor or equal
to zero. This property indicates if a activity is critical:
\code{TRUE} indicates it is critical;
\code{FALSE} indicates it is not critical.
\item \strong{ES:} Early Start: is the earliest start period an activity can begin
after its predecessors without violating precedence relation.
\item \strong{EF:} Early Finish: is the early start plus activity duration.
\item \strong{LS:} Late Start: is the late finish minus activity duration.
\item \strong{LF:} Late Finish: is the latest finish an activity can finish
before their successors without violating precedence relation.
\item \strong{total_float:} It is the amount of period an activity can be
delayed without violating the project duration. Its formula is:
LS - ES or LF - EF.
\item \strong{free_float:} It is the amount of period an activity can be
delayed without violating the start time of the successors activities.
\item \strong{progr_level:} Progressive level is the rank of activities counted
from begin. The level of the activities that don't have predecessor is one;
the level of the other activities, is one plus the maximal level of
their predecessor.
\item \strong{regr_level:} Regressive level is the rank of activities counted
from the end. The level of the activities that don't have successor is the
maximal progressive level; the level of the other activities,
is one minus the minimal level of their successor.
\item \strong{topo_float:} It is the difference between progressive level
and regressive level.
\item Usage: \code{Schedule$activities}
}
}
}
\examples{
# Create a schedule
schedule <- Schedule$new()
schedule$title <- "Fictitious Project Example"
schedule$reference <- "VANHOUCKE, Mario. Measuring time:
  improving project performance using earned value management.
  Gent: Springer, 2009, p. 18"

schedule$has_any_activity  # FALSE
schedule$nr_activities     # 0

# Add activities and relations to it.
schedule$add_act_rel(  1, "a1" , 0, c(2,3,4))
schedule$add_act_rel(  2, "a2" , 4, c(5))
schedule$add_act_rel(  3, "a3" , 9, c(10))
schedule$add_act_rel(  4, "a4" , 1, c(6))
schedule$add_act_rel(  5, "a5" , 4, c(9))
schedule$add_act_rel(  6, "a6" , 5, c(7))
schedule$add_act_rel(  7, "a7" , 1, c(8,11))
schedule$add_act_rel(  8, "a8" , 7, c(12))
schedule$add_act_rel(  9, "a9" , 8, c(12))
schedule$add_act_rel( 10, "a10", 3, c(12))
schedule$add_act_rel( 11, "a11", 3, c(12))
schedule$add_act_rel( 12, "a12", 0)

schedule$has_any_activity  # TRUE
schedule$nr_activities     # 12

schedule$get_activity(10)
schedule$activities
}
\seealso{
\link{criticalpath}, \link{Schedule}.
}
\author{
Rubens Jose Rosa (rubens@rubensjoserosa.com),
Marcos dos Santos, Thiago Marques
\itemize{
\item Usage: \code{Schedule$get_activity(id)}
}
}
