% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_successors.R
\name{all_successors}
\alias{all_successors}
\title{List All Successors}
\usage{
all_successors(schedule, id, ign_to)
}
\arguments{
\item{schedule}{A schedule}

\item{id}{Activity id to listed}

\item{ign_to}{An arc to be ignored: id -> ign_to}
}
\value{
A vector
}
\description{
List all successors from an activity: direct or indirect successors.
}
\examples{
schedule <- make_empty_schedule(
  "Fictitious Project Example",
  "VANHOUCKE, Mario. Measuring time:
    improving project performance using earned value management.
    Gent: Springer, 2009, p. 18"
)
schedule <- add_act_rel(schedule,  1, "a1" , 0, c(2,3,4))
schedule <- add_act_rel(schedule,  2, "a2" , 4, c(5))
schedule <- add_act_rel(schedule,  3, "a3" , 9, c(10))
schedule <- add_act_rel(schedule,  4, "a4" , 1, c(6))
schedule <- add_act_rel(schedule,  5, "a5" , 4, c(9))
schedule <- add_act_rel(schedule,  6, "a6" , 5, c(7))
schedule <- add_act_rel(schedule,  7, "a7" , 1, c(8,11))
schedule <- add_act_rel(schedule,  8, "a8" , 7, c(12))
schedule <- add_act_rel(schedule,  9, "a9" , 8, c(12))
schedule <- add_act_rel(schedule, 10, "a10", 3, c(12))
schedule <- add_act_rel(schedule, 11, "a11", 3, c(12))
schedule <- add_act_rel(schedule, 12, "a12", 0)
message("Schedule info:")
schedule$info
message("Activities info:")
schedule$activities

all_predecessors(schedule, 6)

}
\references{
DOCUMENTAÇÃO SOBRE O MÉTODO: livros, artigos ou sites.
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
