% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{from_data_frame}
\alias{from_data_frame}
\alias{schedule_from_data_frame}
\alias{as_data_frame}
\alias{relations_as_data_frame}
\title{Create a schedule object from data frame or vice-versa}
\arguments{
\item{activities}{data frame with activities.}

\item{relations}{data frame with precedence relations between activities.}

\item{title}{A project title for identification.}

\item{reference}{A reference from project origin,
for example, a book, a paper, a corporation, or nothing.}

\item{what}{Character constant, whether to return info about
"activities", "relations" or "both". The default is "activities".}
}
\value{
A Schedule object with an activities and relations,
and either a data frame or a list two data frames named
\code{activites} and \code{relations}.
}
\description{
The method \code{from_data_frame} creates an schedule object
from one or two data frames, one containing activities lists and the order the
precedence relations between activities. After creation, it is applied
the Critical Path Method (CPM).
The method \code{as_data_frame} does the opposite: it returns the activities
data frame, the relation data frame or both.
}
\examples{
activities <- data.frame(
  id        = 1:17,
  name      = paste("a", as.character(1:17), sep=""),
  duration  = c(1,2,2,4,3,3,3,2,1,1,2,1,1,1,1,2,1)
)
relations <- data.frame(
  from = c(1, 1, 2, 2, 2, 3, 3, 3,  3,  4,  5,  6,  7,
  8,  9, 10, 11, 11, 12, 12, 13, 13, 14, 14, 15, 15),
  to   = c(2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 11, 11, 12,
  13, 14, 15, 16, 17, 16, 17, 16, 17, 16, 17, 16, 17)
)
schedule <- Schedule$new()$from_data_frame(
  activities,
  relations,
  "Project 1: Cost Information System",
  "VANHOUCKE, Mario. Integrated project management and control:
  first comes the theory, then the practice.
  Gent: Springer, 2014, p. 6"
)
schedule$as_data_frame()
schedule$as_data_frame("activities")
schedule$as_data_frame("relations")
schedule$as_data_frame("both")
}
\seealso{
\link{criticalpath}, \link{Schedule}.
}
\author{
Rubens Jose Rosa (rubens@rubensjoserosa.com),
Marcos dos Santos, Thiago Marques
}
