% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_redundant.R
\name{is_redundant}
\alias{is_redundant}
\title{Is an arc redundant}
\usage{
is_redundant(schedule, id_from, id_to)
}
\arguments{
\item{schedule}{A Schedule.}

\item{id_from}{Activity id from.}

\item{id_to}{Activity id to.}
}
\value{
A logical TRUE if an arc is redundant; FALSE if not.
}
\description{
Verifify if an arc between two activities, id_from -> id_to, is redundant.
}
\examples{
schedule <- make_empty_schedule(
  "Fictitious Project Example",
  "VANHOUCKE, Mario. Measuring time:
  improving project performance using earned value management.
  Gent: Springer, 2009, p. 18"
)

schedule <- add_act_rel(schedule,  1, "a1" , 0, c(2,3,4,8))
schedule <- add_act_rel(schedule,  2, "a2" , 4, c(5))
schedule <- add_act_rel(schedule,  3, "a3" , 9, c(10))
schedule <- add_act_rel(schedule,  4, "a4" , 1, c(6,7))
schedule <- add_act_rel(schedule,  5, "a5" , 4, c(9))
schedule <- add_act_rel(schedule,  6, "a6" , 5, c(7,11))
schedule <- add_act_rel(schedule,  7, "a7" , 1, c(8,11))
schedule <- add_act_rel(schedule,  8, "a8" , 7, c(12))
schedule <- add_act_rel(schedule,  9, "a9" , 8, c(12))
schedule <- add_act_rel(schedule, 10, "a10", 3, c(12))
schedule <- add_act_rel(schedule, 11, "a11", 3, c(12))
schedule <- add_act_rel(schedule, 12, "a12", 0)

is_redundant(schedule, 1, 8) # TRUE
is_redundant(schedule, 4, 7) # TRUE
is_redundant(schedule, 6, 11) # TRUE

}
\references{
DOCUMENTAÇÃO SOBRE O MÉTODO: livros, artigos ou sites.
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
