% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{nr_activities}
\alias{nr_activities}
\alias{activities}
\alias{has_any_activity}
\alias{any_activity}
\title{Number of  activities / Has any activity}
\value{
An integer indicating the number of activities
or a logical value TRUE, meaning that the schedule has any activity
or FALSE, for a empty schedule.
}
\description{
Number of activities in a schedule as an integer value,
or a logical value that indicate if the schedule has any activity.
The method Schedule$activities and Schedule$relations
does the opposite: it returns the activities data frame
and the relation data frame from a schedule.
}
\seealso{
\link{criticalpath}, \link{Schedule}.
}
\author{
Rubens Jose Rosa (rubens@rubensjoserosa.com),
Marcos dos Santos, Thiago Marques
}
