% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schedule.R
\name{Schedule}
\alias{Schedule}
\title{R6 Class Representing a Schedule}
\usage{
Schedule$add_activity(id, name="", duration=0)

Schedule$add_relation(from_id, to_id, type, lag)

Schedule$add_act_rel(id, name, duration, relations_id=c(), dir="succ")

Schedule$print()

Schedule$activities_as_data_frame()

Schedule$all_successors(id, ign_to)

Schedule$all_predecessors(id, ign_from)

Schedule$topoi_sp()

Schedule$topoi_ad()

Schedule$topoi_la()

Schedule$topoi_tf()
}
\description{
A schedule has activities and relations data-frames.
With a Schedule, you can calculate the project critical path.
}
\details{
Alguns detalhes a mais.
Vamos ver no que vai dar.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nr_activities}}{An integer value that indicate the quantity of activities in a schedule}

\item{\code{has_any_activity}}{A logical value that indicate if the schedule has some activity.}

\item{\code{nr_relations}}{An integer value that indicate the quantity of relations (arcs) in a schedule}

\item{\code{has_any_relation}}{A logical value that indicate if the schedule has some relation.}

\item{\code{duration}}{An integer value that indicate the duration in periods of a schedule
Make a new empty schedule}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Schedule$new()}}
\item \href{#method-add_activity}{\code{Schedule$add_activity()}}
\item \href{#method-add_relation}{\code{Schedule$add_relation()}}
\item \href{#method-add_act_rel}{\code{Schedule$add_act_rel()}}
\item \href{#method-print}{\code{Schedule$print()}}
\item \href{#method-activities_as_data_frame}{\code{Schedule$activities_as_data_frame()}}
\item \href{#method-all_successors}{\code{Schedule$all_successors()}}
\item \href{#method-all_predecessors}{\code{Schedule$all_predecessors()}}
\item \href{#method-change_durations}{\code{Schedule$change_durations()}}
\item \href{#method-get_activity}{\code{Schedule$get_activity()}}
\item \href{#method-gantt_matrix}{\code{Schedule$gantt_matrix()}}
\item \href{#method-xy_gantt_matrix}{\code{Schedule$xy_gantt_matrix()}}
\item \href{#method-is_redundant}{\code{Schedule$is_redundant()}}
\item \href{#method-topoi_sp}{\code{Schedule$topoi_sp()}}
\item \href{#method-topoi_ad}{\code{Schedule$topoi_ad()}}
\item \href{#method-topoi_la}{\code{Schedule$topoi_la()}}
\item \href{#method-topoi_tf}{\code{Schedule$topoi_tf()}}
\item \href{#method-clone}{\code{Schedule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Make a schedule with no activities and no relations.
The activities and relations will be add in sequence.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$new(title = "", reference = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{A project title for identification.}

\item{\code{reference}}{A reference from project origin,
for example, a book, a paper, a corporation, or nothing.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Schedule object with no activities and no relations
Add an activity to a schedule

Add an activity to a schedule that must exist.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_activity"></a>}}
\if{latex}{\out{\hypertarget{method-add_activity}{}}}
\subsection{Method \code{add_activity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$add_activity(id, name = "", duration = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Activity id. The id will be used to make relation between activities.}

\item{\code{name}}{The name of activity. The default is a empty string.}

\item{\code{duration}}{A number that represents the activity's duration.
It must be equal or grater than zero. The default value is zero.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Schedule object with an activity added and critical path calculated.
Add a relation to a schedule.

Add a relation to a schedule.
If type is not defined, it is assumed to be FS.
If lag is not defined, it is assumed to be zero.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_relation"></a>}}
\if{latex}{\out{\hypertarget{method-add_relation}{}}}
\subsection{Method \code{add_relation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$add_relation(from_id, to_id, type = "FS", lag = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from_id}}{The id of predecessor activity.}

\item{\code{to_id}}{The id of successor activity.}

\item{\code{type}}{Specifies the type of relation between activities.
Its value may be: FS, FF, SS, SF, that means:

FS: Finish-Start relation.
Activity to_id can only start after the finish of activity from_id.

FF: Finish-Finish relation.
Activity to_id must finish together with activity from_id.

SS: Start-Start relation.
Activity to_id must start together with activity from_id.

SF: Start-Finish relation.
Activity to_id must finish when activity from_id starts.}

\item{\code{lag}}{The time period between activities that the successor activity
must be advanced, or lated, after activity from_id.
The must be a integer, less than, equal or greater than zero.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Schedule object with a relation added and critical path calculated.

Add an activity and her relations

Add an activity and her relations to a schedule.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_act_rel"></a>}}
\if{latex}{\out{\hypertarget{method-add_act_rel}{}}}
\subsection{Method \code{add_act_rel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$add_act_rel(id, name, duration, relations_id = c(), dir = "succ")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Activity id. The id will be used to make relation between activities.}

\item{\code{name}}{The name of activity.}

\item{\code{duration}}{A number that represents the activity's duration.
It must be equal or grater than zero.}

\item{\code{relations_id}}{A vector of ids such that will be linked with activity id.
It may be a relations of successor or predecessors.}

\item{\code{dir}}{Direction of relations_id: It may be "succ" or "pred".
If dir="succ" the relations_id will be the successor of tye activity.
If dir="pred" the relations_id will be the predecessor of the activity.

Print a description of the class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Variable parameters

Activities as data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-activities_as_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-activities_as_data_frame}{}}}
\subsection{Method \code{activities_as_data_frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$activities_as_data_frame()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame with all activities listed.
List All Successors

List all successors from an activity: direct or indirect successors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_successors"></a>}}
\if{latex}{\out{\hypertarget{method-all_successors}{}}}
\subsection{Method \code{all_successors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$all_successors(id, ign_to = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Activity id to listed}

\item{\code{ign_to}}{An arc to be ignored: id -> ign_to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector
List All Predecessors

List all predecessors from an activity: direct or indirect predecessors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_predecessors"></a>}}
\if{latex}{\out{\hypertarget{method-all_predecessors}{}}}
\subsection{Method \code{all_predecessors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$all_predecessors(id, ign_from = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Activity id to listed}

\item{\code{ign_from}}{An arc to be ignored: ign_from -> id}

\item{\code{schedule}}{A schedule}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector
Change Activities Duration

Change activities duration and calculate critical path.
This way is faster than creating a new schedule with new durations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-change_durations"></a>}}
\if{latex}{\out{\hypertarget{method-change_durations}{}}}
\subsection{Method \code{change_durations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$change_durations(new_durations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_durations}}{A vector with new duration of activities.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Schedule object.
Get a activity by id
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_activity"></a>}}
\if{latex}{\out{\hypertarget{method-get_activity}{}}}
\subsection{Method \code{get_activity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$get_activity(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{An activity id as defined by the user}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data-frame with one line
Gantt Matrix

Create a matrix that represents a Gantt chart.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gantt_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-gantt_matrix}{}}}
\subsection{Method \code{gantt_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$gantt_matrix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A matrix where 1 indicate that an activity is in executions.
In this matrix, the rows represent activities.
Whereas the columns represent the activity execution period.

Transform a Gantt matrix in x,y coordinates

Transform a Gantt matrix in x,y coordinates.
Each point greater than zero in a Gantt matrix becomes a x, y coordinate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-xy_gantt_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-xy_gantt_matrix}{}}}
\subsection{Method \code{xy_gantt_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$xy_gantt_matrix(gantt = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gantt}}{A Gantt Matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix x, y and weight.

Is an arc redundant

Verify if an arc between two activities, id_from -> id_to, is redundant.
An arc A -> C is redundant if there are A -> B -> C relations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_redundant"></a>}}
\if{latex}{\out{\hypertarget{method-is_redundant}{}}}
\subsection{Method \code{is_redundant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$is_redundant(id_from, id_to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_from}}{Activity id from.}

\item{\code{id_to}}{Activity id to.}

\item{\code{schedule}}{A Schedule.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical TRUE if an arc is redundant; FALSE if not.
SP Serial or Parallel Topological Indicator

It shows the closeness of a network to a serial or parallel graph (1, 20).

As the network becomes serial, the SP increase, until one,
when the network totaly serial
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topoi_sp"></a>}}
\if{latex}{\out{\hypertarget{method-topoi_sp}{}}}
\subsection{Method \code{topoi_sp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$topoi_sp()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A number between 0 and 1, inclusive.
AD Activity Distribution Topological Indicator

Measures the distribution of the activities over the levels (1, 20).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topoi_ad"></a>}}
\if{latex}{\out{\hypertarget{method-topoi_ad}{}}}
\subsection{Method \code{topoi_ad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$topoi_ad()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{schedule}}{A Schedule object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A number between 0 and 1, inclusive.
Length of Arcs Topological Indicator

Measures the presence of long arcs based on the difference between
the progressive level of the end activity and the start node of each arc (relation) (1, 20).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topoi_la"></a>}}
\if{latex}{\out{\hypertarget{method-topoi_la}{}}}
\subsection{Method \code{topoi_la()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$topoi_la()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{schedule}}{A Schedule object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A number between 0 and 1, inclusive.
TF Topological Float Indicator

Measures the topological float of each activity (1, 20).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-topoi_tf"></a>}}
\if{latex}{\out{\hypertarget{method-topoi_tf}{}}}
\subsection{Method \code{topoi_tf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$topoi_tf()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A number between 0 and 1, inclusive.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schedule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
