% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_from_data_frame.R
\name{schedule_from_data_frame}
\alias{schedule_from_data_frame}
\title{Create a schedule from data frame}
\usage{
schedule_from_data_frame(activities, relations, title, reference)
}
\arguments{
\item{activities}{data frame with activities.}

\item{relations}{data frame with precedence relations between activities.}

\item{title}{A project title for identification.}

\item{reference}{A reference from project origin,
for example, a book, a paper, a corporation, or nothing.}
}
\value{
A Schedule object with an activity added and critical path calculated.
}
\description{
Create a schedule from data frames and apply Critical Path Method (CPM)
}
\examples{
activities <- data.frame(
  id        = 1:17,
  name      = paste("a", as.character(1:17), sep=""),
  duration  = c(1,2,2,4,3,3,3,2,1,1,2,1,1,1,1,2,1)
)
relations <- data.frame(
  from = c(1, 1, 2, 2, 2, 3, 3, 3,  3,  4,  5,  6,  7,  8,  9,
           10, 11, 11, 12, 12, 13, 13, 14, 14, 15, 15),
  to   = c(2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 11, 11, 12, 13, 14,
           15, 16, 17, 16, 17, 16, 17, 16, 17, 16, 17)
)
schedule <- schedule_from_data_frame(
  activities,
  relations,
  "Project 1: Cost Information System",
  "VANHOUCKE, Mario. Integrated project management and control:
    first comes the theory, then the practice. Gent: Springer, 2014, p. 6"
)

}
\references{
DOCUMENTAÇÃO SOBRE O MÉTODO: livros, artigos ou sites.
}
\seealso{
Creating a schedule:
\code{\link[=schedule_from_data_frame]{schedule_from_data_frame()}}
\code{\link[=make_empty_schedule]{make_empty_schedule()}}
\code{\link[=add_activity]{add_activity()}}
\code{\link[=add_relation]{add_relation()}}
\code{\link[=add_act_rel]{add_act_rel()}}

Changing the activities duration:
\code{\link[=change_durations]{change_durations()}}

Gantt Matrix:
\code{\link[=gantt_matrix]{gantt_matrix()}}
\code{\link[=xy_gantt_matrix]{xy_gantt_matrix()}}

Topological indicators:
\code{\link[=topoi_sp]{topoi_sp()}}
\code{\link[=topoi_ad]{topoi_ad()}}
\code{\link[=topoi_la]{topoi_la()}}
\code{\link[=topoi_tf]{topoi_tf()}}
}
