% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath.R
\name{topological_indicators}
\alias{topological_indicators}
\alias{topoi_sp}
\alias{topoi_ad}
\alias{topoi_la}
\alias{topoi_tf}
\alias{sp}
\alias{ad}
\alias{la}
\alias{tf}
\alias{topological}
\alias{indicator}
\alias{indicators}
\title{Topological Indicators: SP, AD, LA, TF}
\value{
A number between 0 and 1, inclusive.
}
\description{
\itemize{
\item $topoi_sp()
\item $topoi_ad()
\item $topoi_la()
\item $topoi_tf()
}

Shows information about network structure.
It may be of four type:

\strong{SP Serial or Parallel:}
It shows the closeness of a network to a serial or parallel graph.
As the network becomes serial, the SP increase, until one,
when the network totally serial.

\strong{AD Activity Distribution:}
Measures the distribution of the activities over the levels.
If AD is approximately equal zero, each level has same numbers of activities.
Otherwise, if AD is equal one, the quantity of each level is not
uniformly distributed.

\strong{LA Length of Arcs:}
Measures the presence of long arcs based on the difference between
the progressive level of the end activity and the start node
of each relation.
If LA is approximately equal zero, the progressive level between
activities are as far as possible.
Otherwise, if LA is equal one, the relation distance are one.

\strong{TF Topological Float Indicator:}
Measures the topological float of each activity.
If TF = 0, there is no float between activities.
If TF = 1, there is float between activities
and they be shift without affecting other activities.
}
\seealso{
\link{criticalpath}, \link{Schedule}.
}
\author{
Rubens Jose Rosa (rubens@rubensjoserosa.com),
Marcos dos Santos, Thiago Marques
}
