% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_organization.R
\name{topological_organization}
\alias{topological_organization}
\title{Topological Organization}
\usage{
topological_organization(schedule)
}
\arguments{
\item{schedule}{}
}
\value{
schedule
}
\description{
This function is for internal use! It calculates topological definitions
that is necessary before critical path method.
}
\references{
Mario Vanhoucke

Rubens Jose Rosa

PMBOK
}
